cw_direction_sign int -> int8_t

+++ src/core1.cpp
@@ -26,7 +26,7 @@
 
 const int full_steps_per_turn = 200;
 const float deg_per_full_step = 1.8;
-const int cw_direction_sign = -1;
+const int8_t cw_direction_sign = -1;
 
 HR4988 stepper_motor = HR4988 (
     STEP_PIN, DIRECTION_PIN,
